Made the complete tasks guard generic, so it can be used by other flows (e.g. claim flow)
