+++ src/main/routes/features/response/checkAnswersController.ts
@@ -1,5 +1,9 @@
 import * as express from 'express';
-import {CONFIRMATION_URL, RESPONSE_CHECK_ANSWERS_URL} from '../../urls';
+import {
+  CONFIRMATION_URL,
+  RESPONSE_CHECK_ANSWERS_URL,
+  RESPONSE_INCOMPLETE_SUBMISSION_URL,
+} from '../../urls';
 import {
   getStatementOfTruth,
   getSummarySections,
@@ -11,7 +15,7 @@ import {getCaseDataFromStore} from '../../../modules/draft-store/draftStoreServi
 import {StatementOfTruthForm} from '../../../common/form/models/statementOfTruth/statementOfTruthForm';
 import {Claim} from '../../../common/models/claim';
 import {constructResponseUrlWithIdParams} from '../../../common/utils/urlFormatter';
-import {AllResponseTasksCompletedGuard} from '../../../routes/features/response/guards/allResponseTasksCompletedGuard';
+import {AllResponseTasksCompletedGuard} from '../../guards/allResponseTasksCompletedGuard';
 
 const {Logger} = require('@hmcts/nodejs-logging');
 const logger = Logger.getLogger('checkAnswersController');
@@ -31,12 +35,12 @@ function renderView(req: express.Request, res: express.Response, form: GenericFo
 
 
 checkAnswersController.get(RESPONSE_CHECK_ANSWERS_URL,
-  AllResponseTasksCompletedGuard.apply,
-  async (req, res) => {
+  AllResponseTasksCompletedGuard.apply(RESPONSE_INCOMPLETE_SUBMISSION_URL),
+  async (req: express.Request, res: express.Response) => {
     try {
       const claim = await getCaseDataFromStore(req.params.id);
       const form = new GenericForm(getStatementOfTruth(claim));
-      await renderView(req, res, form, claim);
+      renderView(req, res, form, claim);
     } catch (error) {
       logger.error(error);
       res.status(500).send({error: error.message});
@@ -49,7 +53,7 @@ checkAnswersController.post(RESPONSE_CHECK_ANSWERS_URL, async (req: express.Requ
     await form.validate();
     if (form.hasErrors()) {
       const claim = await getCaseDataFromStore(req.params.id);
-      await renderView(req, res, form, claim);
+      renderView(req, res, form, claim);
     } else {
       await saveStatementOfTruth(req.params.id, form.model);
       res.redirect(constructResponseUrlWithIdParams(req.params.id, CONFIRMATION_URL));
